Add ContentProxy a.k.a (try) to add Scrambled Exif as a file chooser.
AndroidManifest.xml: add IntentFilter for GET_CONTENT.
If you choose the file manager (documentsui), e.g. when asked to upload
a file to a website you can now choose Scrambled Exif as a source in
the burger menu.

ContentProxyActivity: responds to the GET_CONTENT defined in the
AndroidManifest.xml.
It allows to pick *one* image from the gallery and send a scrambled copy
to the calling app using a "result" intent.

I don't seem to be able to return multiple uris using clipdata
(for >= JELLY_BEAN_MR2).

It would be nice, if it would get shown in the first "open with" intent.

Partially fixes #7.