Add Panic Button settings and respond to an "EDIT" action from Ripple (actually, just added ACTION_CONNECT and ACTION_DISCONNECT as intent-filters for the SettingsActivity - this is probably not the nicest way to do it).
Improve fix for #32.

Note: There is a way to actually CONNECT and DISCONNECT. Panickit
(https://guardianproject.info/code/panickit/) could help here, but I believe
the current implementation might suffice. The problem could be that we don't
check at all which app is triggering the panic, so any app could trigger it.
Because of Scrambled Exif's "bridge" nature, it doesn't need to store anything
and the only thing that might get lost are the (very few) settings.