Remove READ_EXTERNAL_STORAGE permission.
Instead of removing all the code 'isPermissionGranted' always returns
true. Like this, I can revert this change more easily, should this
lead to unexpected errors.

TODO: remove all the code that calls 'isPermissionGranted' and the
corresponding UI in MainActivity (show a button and explanation
or not) if this works out nicely.

Fixes #41.