Remove READ_EXTERNAL_STORAGE permission for API>=30.
Instead of removing all the code 'isPermissionGranted' returns
true if API>=30. Like this, I can revert this change more easily, should this
lead to unexpected errors.

Note: I already did this for *all* SDK levels and it was a huge
PITA as a lot of apps don't grant the permission to the Uris correctly
and Scrambled Exif crashes while trying to read them...

TODO: remove all the code that calls 'isPermissionGranted' and the
corresponding UI in MainActivity (show a button and explanation
or not) if this works out nicely.

See #41.