Revert "Remove READ_EXTERNAL_STORAGE permission for API>=30."
This reverts commit 32d09b48b9ec043986456add786d5950e4681f68.

Re-add READ_EXTERNAL_STORAGE for API>=30. On some devices, it's
still crashing. Hopefully, this'll help. I don't really like to
ask for a permission we shouldn't need, but it actually depends
on which the other apps share the images correctly (with
a grantUriPermission), which a lot still don't do.

See #41 and #48.