Fix Dark Mode - Update app/build.gradle
this marks the fixing of Dark mode. App should now set Dark Mode
according to how the System Mode is set.

A theme option in settings wasn't added since the app isn't to be
opened often anyways. Tried adding but didn't turn out to be that
simple. Might come back to it at a later point